Reorganized firebase services
