+++ frontend/s_a_m_e/lib/account/edit_symptoms.dart
@@ -1,41 +1,38 @@
 import 'package:flutter/material.dart';
-import 'package:s_a_m_e/account/account.dart';
 import 'package:s_a_m_e/account/profilepicture.dart';
 import 'package:s_a_m_e/admin/add_symptom.dart';
-import 'package:s_a_m_e/admin/adminrequest.dart';
 import 'package:s_a_m_e/admin/delete_symptom.dart';
 import 'package:s_a_m_e/admin/update_symptom.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/account/login.dart';
-import 'package:s_a_m_e/admin/view_accounts.dart';
 import 'package:s_a_m_e/user/symptom_list.dart';
 
 class SymptomEdit extends StatelessWidget {
   //const Admin({super.key});
 
-@override
-Widget build(BuildContext context) {
-  return Scaffold(
-    appBar: AppBar(
-      title: const Text('S.A.M.E'),
-      actions: [ProfilePicturePage()],
-    ),
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('S.A.M.E'),
+        actions: [ProfilePicturePage()],
+      ),
       body: Align(
         alignment: Alignment.topCenter,
         child: Column(
           children: <Widget>[
             const SizedBox(height: 25),
-            const Text('Hello Admin User!', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0)),
+            const Text('Hello Admin User!',
+                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0)),
             const SizedBox(height: 10),
-            
-            
             const SizedBox(height: 10),
             ElevatedButton(
               style: const ButtonStyle(
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Add Sypmtom', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Add Sypmtom',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -51,7 +48,9 @@ Widget build(BuildContext context) {
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Delete Sypmtom', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Delete Sypmtom',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -67,7 +66,9 @@ Widget build(BuildContext context) {
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Update Sypmtom', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Update Sypmtom',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -83,7 +84,9 @@ Widget build(BuildContext context) {
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Sypmtoms List', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Sypmtoms List',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -92,9 +95,7 @@ Widget build(BuildContext context) {
                   ),
                 );
               },
-              
             ),
-            
           ],
         ),
       ),
@@ -102,4 +103,5 @@ Widget build(BuildContext context) {
   }
 }
 
-String adminDisclaimer = "This application is for general informational purposes only and is not intended to be a substitute for professional medical advice. The information provided should not be considered as medical advice, and we do not guarantee its accuracy.\nOur application is a guideline, and individual cases may vary.\n\nWe are not liable for any loss or damage arising from the use of this information. Consult with qualified healthcare professionals for advice tailored to your specific circumstances.\n\nBy using this application, you agree to these terms. S.A.M.E. is not responsible for errors or omissions. Use this application responsibly and in accordance with applicable laws.\n";
+String adminDisclaimer =
+    "This application is for general informational purposes only and is not intended to be a substitute for professional medical advice. The information provided should not be considered as medical advice, and we do not guarantee its accuracy.\nOur application is a guideline, and individual cases may vary.\n\nWe are not liable for any loss or damage arising from the use of this information. Consult with qualified healthcare professionals for advice tailored to your specific circumstances.\n\nBy using this application, you agree to these terms. S.A.M.E. is not responsible for errors or omissions. Use this application responsibly and in accordance with applicable laws.\n";
