+++ frontend/s_a_m_e/lib/account/login.dart
@@ -7,6 +7,7 @@ import 'package:s_a_m_e/colors.dart';
 import 'package:s_a_m_e/account/signup.dart';
 import 'package:s_a_m_e/user/user_home.dart';
 import 'package:s_a_m_e/firebase/firebase_service.dart';
+import '../firebase/models.dart';
 
 class Login extends StatefulWidget {
   const Login({Key? key}) : super(key: key);
@@ -30,43 +31,41 @@ class LoginState extends State<Login> {
   }
 
   Future<void> _signIn() async {
-  try {
-    UserCredential userCredential = await _auth.signInWithEmailAndPassword(
-      email: _emailController.text.trim(),
-      password: _passwordController.text,
-    );
+    try {
+      UserCredential userCredential = await _auth.signInWithEmailAndPassword(
+        email: _emailController.text.trim(),
+        password: _passwordController.text,
+      );
 
-    if (userCredential.user != null) {
-      String uid = userCredential.user!.uid;
+      if (userCredential.user != null) {
+        String uid = userCredential.user!.uid;
 
-      UserClass? userData = await FirebaseService().getUser(uid);
+        UserClass? userData = await FirebaseService().getUser(uid);
 
-      if (userData != null) {
-        String role = userData.role;
-        print(role);
+        if (userData != null) {
+          String role = userData.role;
+          print(role);
 
-        if (role == 'user') {
-          Navigator.of(context).push(
-            MaterialPageRoute(builder: (context) => const UserHome()),
-          );
-        } else if (role == 'admin') {
-          Navigator.of(context).push(
-            MaterialPageRoute(builder: (context) => const Admin()),
-          );
+          if (role == 'user') {
+            Navigator.of(context).push(
+              MaterialPageRoute(builder: (context) => const UserHome()),
+            );
+          } else if (role == 'admin') {
+            Navigator.of(context).push(
+              MaterialPageRoute(builder: (context) => const Admin()),
+            );
+          }
+        } else {
+          print('Error: User data not found in the database.');
         }
       } else {
-        print('Error: User data not found in the database.');
+        print('Error: User is null after sign-in.');
       }
-    } else {
-      print('Error: User is null after sign-in.');
+    } catch (e) {
+      print('Error during sign-in: $e');
+      // Handle exceptions
     }
-  } catch (e) {
-    print('Error during sign-in: $e');
-    // Handle exceptions
   }
-}
-
-
 
   void _showDisclaimerDialog(BuildContext context) {
     bool checkboxValue = false;
@@ -90,11 +89,12 @@ class LoginState extends State<Login> {
                         color: Colors.black,
                       ),
                     ),
-                    const SizedBox(height: 10), 
+                    const SizedBox(height: 10),
                     Row(
                       children: <Widget>[
                         Checkbox(
-                          fillColor: MaterialStateProperty.resolveWith((states) {
+                          fillColor:
+                              MaterialStateProperty.resolveWith((states) {
                             if (!states.contains(MaterialState.selected)) {
                               return Colors.transparent;
                             }
@@ -115,7 +115,7 @@ class LoginState extends State<Login> {
                           style: TextStyle(
                             fontFamily: "PT Serif",
                             fontSize: 14.0,
-                            color: Colors.black, 
+                            color: Colors.black,
                           ),
                         ),
                       ],
@@ -132,7 +132,7 @@ class LoginState extends State<Login> {
                       style: TextStyle(
                         fontFamily: "PT Serif",
                         fontSize: 16.0,
-                        color: Colors.black, 
+                        color: Colors.black,
                       )),
                 ),
                 ElevatedButton(
@@ -185,31 +185,28 @@ class LoginState extends State<Login> {
           child: Column(
             children: <Widget>[
               const SizedBox(height: 15),
-              const Image(
-                height: 220,
-                image: AssetImage('assets/logo.png')
-              ),
-              const SizedBox(height: 20), 
-              const Text('Welcome back!', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 30.0)),
-              const Text('Login with your credentials below', style: TextStyle(fontSize: 14.0)),
+              const Image(height: 220, image: AssetImage('assets/logo.png')),
+              const SizedBox(height: 20),
+              const Text('Welcome back!',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 30.0)),
+              const Text('Login with your credentials below',
+                  style: TextStyle(fontSize: 14.0)),
               const SizedBox(height: 40),
               TextField(
                 controller: _emailController,
                 decoration: const InputDecoration(
-                  contentPadding: EdgeInsets.all(20.0),
-                  labelText: 'Email',
-                  labelStyle: TextStyle(color: navy),
-                  filled: true,
-                  fillColor: boxinsides,
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.all(Radius.circular(40.0)),
-                    borderSide: BorderSide(color: boxinsides)
-                  ),
-                  focusedBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.all(Radius.circular(40.0)),
-                    borderSide: BorderSide(color: boxinsides)
-                  )
-                ),
+                    contentPadding: EdgeInsets.all(20.0),
+                    labelText: 'Email',
+                    labelStyle: TextStyle(color: navy),
+                    filled: true,
+                    fillColor: boxinsides,
+                    enabledBorder: OutlineInputBorder(
+                        borderRadius: BorderRadius.all(Radius.circular(40.0)),
+                        borderSide: BorderSide(color: boxinsides)),
+                    focusedBorder: OutlineInputBorder(
+                        borderRadius: BorderRadius.all(Radius.circular(40.0)),
+                        borderSide: BorderSide(color: boxinsides))),
               ),
               const SizedBox(height: 30),
               TextField(
@@ -231,35 +228,35 @@ class LoginState extends State<Login> {
               ),
               const SizedBox(height: 40),
               RichText(
-                text: TextSpan(
+                  text: TextSpan(
                 style: const TextStyle(fontFamily: "PT Serif"),
                 children: <TextSpan>[
                   const TextSpan(
                       text: "Don't have an account?  ",
                       style: TextStyle(color: Colors.black)),
                   TextSpan(
-                    text: "Register here",
-                    style: const TextStyle(color: blue, decoration: TextDecoration.underline),
-                    recognizer: TapGestureRecognizer()
-                      ..onTap = () {
-                        _showDisclaimerDialog(context);
-                      } 
-                  ),
+                      text: "Register here",
+                      style: const TextStyle(
+                          color: blue, decoration: TextDecoration.underline),
+                      recognizer: TapGestureRecognizer()
+                        ..onTap = () {
+                          _showDisclaimerDialog(context);
+                        }),
                 ],
               )),
               const SizedBox(height: 40),
               SizedBox(
-                    width: MediaQuery.of(context).size.width,
-                    height: 50,
-                    child: ElevatedButton(
-                        style: const ButtonStyle(
-                          foregroundColor: MaterialStatePropertyAll<Color>(white),
-                          backgroundColor: MaterialStatePropertyAll<Color>(navy),
-                        ),
-                        onPressed: _signIn,
-                        child: const Text('Sign In',
-                            style: TextStyle(
-                                fontWeight: FontWeight.bold, fontSize: 24.0)))),
+                  width: MediaQuery.of(context).size.width,
+                  height: 50,
+                  child: ElevatedButton(
+                      style: const ButtonStyle(
+                        foregroundColor: MaterialStatePropertyAll<Color>(white),
+                        backgroundColor: MaterialStatePropertyAll<Color>(navy),
+                      ),
+                      onPressed: _signIn,
+                      child: const Text('Sign In',
+                          style: TextStyle(
+                              fontWeight: FontWeight.bold, fontSize: 24.0)))),
               const SizedBox(height: 30),
               // const Text('Forgot password?'),
             ],
