+++ frontend/s_a_m_e/lib/admin/admin_home.dart
@@ -3,100 +3,99 @@ import 'package:s_a_m_e/account/account.dart';
 import 'package:s_a_m_e/account/edit_signs.dart';
 import 'package:s_a_m_e/account/edit_symptoms.dart';
 import 'package:s_a_m_e/account/profilepicture.dart';
-import 'package:s_a_m_e/admin/add_category.dart';
-import 'package:s_a_m_e/admin/add_sign.dart';
 import 'package:s_a_m_e/admin/adminrequest.dart';
-import 'package:s_a_m_e/admin/delete_sign.dart';
 import 'package:s_a_m_e/admin/edit_categories.dart';
-import 'package:s_a_m_e/admin/update_sign.dart';
 import 'package:s_a_m_e/user/diagnosis_list.dart';
-import 'package:s_a_m_e/user/categories_list.dart';
 import 'package:s_a_m_e/colors.dart';
-import 'package:s_a_m_e/admin/add_symptom.dart';
-import 'package:s_a_m_e/admin/delete_category.dart';
-import 'package:s_a_m_e/admin/delete_symptom.dart';
 import 'package:s_a_m_e/admin/edit_diagnoses.dart';
 import 'package:s_a_m_e/account/login.dart';
-import 'package:s_a_m_e/user/sign_list.dart';
-import 'package:s_a_m_e/user/symptom_list.dart';
-import 'package:s_a_m_e/admin/update_category.dart';
-import 'package:s_a_m_e/admin/update_symptom.dart';
 import 'package:s_a_m_e/admin/view_accounts.dart';
 import 'package:s_a_m_e/userflow/select_symptoms.dart';
 
 class Admin extends StatelessWidget {
   const Admin({super.key});
 
-@override
-Widget build(BuildContext context) {
-  return Scaffold(
-    appBar: AppBar(
-      title: const Text('S.A.M.E'),
-      actions: [ProfilePicturePage()],
-    ),
-    drawer: Drawer(
-      backgroundColor: background,
-      child: ListView(
-        children: <Widget>[
-          ListTile(
-            title: const Text('Manage Account'),
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => const ManageAccountPage(),
-                ),
-              );
-            },
-          ),
-          ListTile(
-            title: const Text('View All Users'),
-            onTap: () {
-              Navigator.push(context, MaterialPageRoute(builder: (context) => const ViewAccounts(),));
-            },
-          ),
-          ListTile(
-            title: const Text('Admin Requests'),
-            onTap: () {
-              Navigator.push(context, MaterialPageRoute(builder: (context) => const AdminRequestPage(),));
-            },
-          ),
-          // ListTile(
-          //   title: const Text('Edit Diagnosis'),
-          //   onTap: () {
-          //     Navigator.push(context, MaterialPageRoute(builder: (context) => const EditDiagnosisPage(),));
-          //   },
-          // ),
-          ListTile(
-            title: const Text('Sign Out'),
-            onTap: () {
-              Navigator.of(context).popUntil((route) => route.isFirst);
-              Navigator.pushReplacement(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => const Login(),
-                ),
-              );
-            },
-          ),
-        ],
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('S.A.M.E'),
+        actions: [ProfilePicturePage()],
       ),
-    ),
+      drawer: Drawer(
+        backgroundColor: background,
+        child: ListView(
+          children: <Widget>[
+            ListTile(
+              title: const Text('Manage Account'),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const ManageAccountPage(),
+                  ),
+                );
+              },
+            ),
+            ListTile(
+              title: const Text('View All Users'),
+              onTap: () {
+                Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => const ViewAccounts(),
+                    ));
+              },
+            ),
+            ListTile(
+              title: const Text('Admin Requests'),
+              onTap: () {
+                Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => const AdminRequestPage(),
+                    ));
+              },
+            ),
+            // ListTile(
+            //   title: const Text('Edit Diagnosis'),
+            //   onTap: () {
+            //     Navigator.push(context, MaterialPageRoute(builder: (context) => const EditDiagnosisPage(),));
+            //   },
+            // ),
+            ListTile(
+              title: const Text('Sign Out'),
+              onTap: () {
+                Navigator.of(context).popUntil((route) => route.isFirst);
+                Navigator.pushReplacement(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const Login(),
+                  ),
+                );
+              },
+            ),
+          ],
+        ),
+      ),
       body: Align(
         alignment: Alignment.topCenter,
         child: Column(
           children: <Widget>[
             const SizedBox(height: 25),
-            const Text('Hello Admin User!', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0)),
+            const Text('Hello Admin User!',
+                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0)),
             const SizedBox(height: 10),
-            
+
             const SizedBox(height: 10),
             ElevatedButton(
               style: const ButtonStyle(
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Diagnoses', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Diagnoses',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -112,7 +111,9 @@ Widget build(BuildContext context) {
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Signs', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Signs',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -122,7 +123,7 @@ Widget build(BuildContext context) {
                 );
               },
             ),
-            
+
             //const SizedBox(height: 10),
             const SizedBox(height: 10),
             ElevatedButton(
@@ -130,7 +131,9 @@ Widget build(BuildContext context) {
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Symptoms', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Symptoms',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -140,15 +143,16 @@ Widget build(BuildContext context) {
                 );
               },
             ),
-            
-            
+
             const SizedBox(height: 10),
             ElevatedButton(
               style: const ButtonStyle(
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Categories', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Categories',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -159,14 +163,15 @@ Widget build(BuildContext context) {
               },
             ),
 
-            
             const SizedBox(height: 10),
             ElevatedButton(
               style: const ButtonStyle(
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Get Potential Diagnosis', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Get Potential Diagnosis',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -181,7 +186,9 @@ Widget build(BuildContext context) {
                 foregroundColor: MaterialStatePropertyAll<Color>(white),
                 backgroundColor: MaterialStatePropertyAll<Color>(navy),
               ),
-              child: const Text('Diagnosis List', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
+              child: const Text('Diagnosis List',
+                  style:
+                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16.0)),
               onPressed: () {
                 Navigator.push(
                   context,
@@ -197,4 +204,5 @@ Widget build(BuildContext context) {
   }
 }
 
-String adminDisclaimer = "This application is for general informational purposes only and is not intended to be a substitute for professional medical advice. The information provided should not be considered as medical advice, and we do not guarantee its accuracy.\nOur application is a guideline, and individual cases may vary.\n\nWe are not liable for any loss or damage arising from the use of this information. Consult with qualified healthcare professionals for advice tailored to your specific circumstances.\n\nBy using this application, you agree to these terms. S.A.M.E. is not responsible for errors or omissions. Use this application responsibly and in accordance with applicable laws.\n";
+String adminDisclaimer =
+    "This application is for general informational purposes only and is not intended to be a substitute for professional medical advice. The information provided should not be considered as medical advice, and we do not guarantee its accuracy.\nOur application is a guideline, and individual cases may vary.\n\nWe are not liable for any loss or damage arising from the use of this information. Consult with qualified healthcare professionals for advice tailored to your specific circumstances.\n\nBy using this application, you agree to these terms. S.A.M.E. is not responsible for errors or omissions. Use this application responsibly and in accordance with applicable laws.\n";
